feat: reorganize authentication DTOs into auth package for better structure
