+++ services/auth-service/src/main/java/com/academiaconnect/auth/authservice/application/dto/auth/LoginRequest.java
@@ -1,4 +1,4 @@
-package com.academiaconnect.auth.authservice.application.dto;
+package com.academiaconnect.auth.authservice.application.dto.auth;
 
 import com.fasterxml.jackson.annotation.JsonAlias;
 import jakarta.validation.constraints.NotBlank;
