+++ services/auth-service/src/main/java/com/academiaconnect/auth/authservice/application/dto/auth/PasswordUpdateDto.java
@@ -1,4 +1,4 @@
-package com.academiaconnect.auth.authservice.application.dto;
+package com.academiaconnect.auth.authservice.application.dto.auth;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.Size;
