+++ services/auth-service/src/main/java/com/academiaconnect/auth/authservice/application/dto/auth/ResetPasswordRequest.java
@@ -1,4 +1,4 @@
-package com.academiaconnect.auth.authservice.application.dto;
+package com.academiaconnect.auth.authservice.application.dto.auth;
 
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
