+++ services/auth-service/src/main/java/com/academiaconnect/auth/authservice/application/dto/auth/TokenResponse.java
@@ -1,4 +1,4 @@
-package com.academiaconnect.auth.authservice.application.dto;
+package com.academiaconnect.auth.authservice.application.dto.auth;
 
 import lombok.*;
